pyproject.toml
[project]
name = "ds6050-project"
version = "0.1.0"
description = "yo yo Add your description here yo"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datasets>=3.0,<4.0",
    "evaluate>=0.4.6",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "seqeval>=1.2.2",
    "pytorch-crf>=0.7.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
]
